refactor(cli): decompose index.ts monolith into modular command modules
Problem:
index.ts is 2818 lines - a monolithic entry point mixing CLI setup,
15+ command definitions, interactive mode, workflow logic, and service
instantiation. This makes testing, maintenance, and onboarding
difficult. Individual command groups cannot be unit tested in isolation.

Solution:

  New: CLIDependencies (src/cli/types.ts)
  - Interface centralizing shared dependencies: program, ui, t, version,
    defaultModel
  - Enables clean dependency injection into command modules
  - Promotes testability: modules can receive mock dependencies

  New: skillCommands.ts (src/cli/commands/skillCommands.ts)
  - registerSkillCommands(deps: CLIDependencies): void
  - Extracted 5 subcommands: skill init, skill fill, skill list,
    skill export, skill create
  - Each subcommand preserves original functionality exactly
  - Dynamic imports maintained for code splitting

  New: workflowCommands.ts (src/cli/commands/workflowCommands.ts)
  - registerWorkflowCommands(deps: CLIDependencies): void
  - Extracted 6 subcommands: workflow init, workflow status,
    workflow advance, workflow handoff, workflow collaborate, workflow role
  - Creates WorkflowServiceDependencies from CLIDependencies
  - Preserves all PHASE_NAMES_PT/EN, ROLE_DISPLAY_NAMES references

  New: commands/index.ts (src/cli/commands/index.ts)
  - Barrel export for registerSkillCommands, registerWorkflowCommands

  Modified: index.ts (src/index.ts)
  - Import registerSkillCommands, registerWorkflowCommands from cli/commands
  - Replaced ~350 lines of inline skill/workflow command registration
    with 4-line import + function calls
  - Added inline WorkflowServiceDependencies for interactive mode
    (runInteractiveWorkflow function)
  - Reduced from 2818 to 2478 lines (340 lines extracted)

  Modified: mcpServer.ts (src/services/mcp/mcpServer.ts)
  - Import PathValidator from utils/pathSecurity
  - Import ContextCache from services/semantic/contextCache
  - Initialize PathValidator with workspace root in constructor
  - Initialize ContextCache as class field
  - Validate file paths in wrapWithActionLogging before tool execution
  - Use cache-first pattern in registerResources() context handler

Impact:
- index.ts reduced from 2818 to 2478 lines (12% reduction)
- Established modular pattern for continued decomposition
- Each command group can now be tested in isolation
- MCP server hardened with security validation and caching
- Zero breaking changes: public API surface unchanged
- Pattern ready for extracting remaining commands in future PRs
