+++ src/cli/commands/index.ts
@@ -1,0 +1,8 @@
+/**
+ * CLI Commands Index
+ *
+ * Re-exports all extracted command registration functions.
+ */
+
+export { registerSkillCommands } from './skillCommands';
+export { registerWorkflowCommands } from './workflowCommands';
